Make database extension code object oriented
