+++ libraries/Error.class.php
@@ -306,7 +306,6 @@ class PMA_Error extends PMA_Message
             'mysqli_connect',
             'mysqli_real_connect',
             'PMA_DBI_connect',
-            'PMA_DBI_realConnect',
         );
 
         if (in_array($function, $include_functions)) {
