+++ test/libraries/database_interface_test.php
@@ -39,7 +39,7 @@ class PMA_DBI_Test extends PHPUnit_Framework_TestCase
      */
     function testQuery()
     {
-        $this->assertEquals(0, PMA_DBI_realQuery('SELECT 1'));
+        $this->assertEquals(0, PMA_DBI_tryQuery('SELECT 1'));
     }
 
     /**
